chore: setup the editorconfig
Summary:

Now formatting of the cs files are formatted with `dotnet format`. There are
still config files that don't, we will need to watch when adding and editing
them. For now they are all lf, so we will see the edits in the diffs.

Test Plan:

CI, this in the cs script